Rename pytx3 to threatexchange
Moving to the threatexchange name for the python CLI and library
package.
