+++ python-threatexchange/.gitignore
@@ -1,5 +1,5 @@
 __pycache__
 build/
 dist/
-pytx3.egg-info/
+threatexchange.egg-info/
 .eggs/
