+++ python-threatexchange/DESCRIPTION.rst
@@ -1,0 +1,1 @@
+threatexchange is a Python Library and CLI tool for interfacing with Facebook's ThreatExchange.
