+++ python-threatexchange/tests/test_common.py
@@ -2,9 +2,9 @@
 
 import unittest
 
-import pytx3.common
+import threatexchange.common
 
 
 class TestCommon(unittest.TestCase):
     def test_camel_case_to_underscore(self):
-        assert pytx3.common.camel_case_to_underscore("AbcXyz") == "abc_xyz"
+        assert threatexchange.common.camel_case_to_underscore("AbcXyz") == "abc_xyz"
