+++ python-threatexchange/threatexchange/cli/command_base.py
