+++ python-threatexchange/threatexchange/cli/main.py
