+++ python-threatexchange/threatexchange/common.py
