+++ python-threatexchange/threatexchange/content_type/content_base.py
