+++ python-threatexchange/threatexchange/content_type/photo.py
