+++ python-threatexchange/threatexchange/content_type/text.py
