+++ python-threatexchange/threatexchange/descriptor.py
