+++ python-threatexchange/threatexchange/indicator.py
