+++ python-threatexchange/threatexchange/signal_type/signal_base.py
