+++ python-threatexchange/threatexchange/signal_type/video_tmk_pdqf.py
