--- pytx3/DESCRIPTION.rst
@@ -1,1 +1,0 @@
-pytx3 (pie-tex) is a Python Library for interfacing with Facebook's ThreatExchange.
